+++ field_history/middleware.py
@@ -1,0 +1,7 @@
+from .tracker import FieldHistoryTracker
+
+
+class FieldHistoryMiddleware(object):
+
+    def process_request(self, request):
+        FieldHistoryTracker.thread.request = request